+++ space-d/src/main/java/com/dnd/spaced/core/auth/infrastructure/jwt/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.auth.infrastructure.exception;
+package com.dnd.spaced.core.auth.infrastructure.jwt.exception;
 
 import com.dnd.spaced.global.exception.base.AuthClientException;
 import com.dnd.spaced.global.exception.code.AuthErrorCode;
