+++ space-d/src/test/java/com/dnd/spaced/core/auth/application/TokenServiceTest.java
@@ -13,8 +13,8 @@ import com.dnd.spaced.core.auth.domain.TokenEncoder;
 import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import com.dnd.spaced.core.auth.domain.repository.BlacklistTokenRepository;
 import com.dnd.spaced.core.auth.domain.repository.RefreshTokenRotationRepository;
-import com.dnd.spaced.core.auth.infrastructure.JwtEncoder;
-import com.dnd.spaced.core.auth.infrastructure.exception.InvalidTokenException;
+import com.dnd.spaced.core.auth.infrastructure.jwt.JwtEncoder;
+import com.dnd.spaced.core.auth.infrastructure.jwt.exception.InvalidTokenException;
 import com.dnd.spaced.fixture.LocalDateTimeFixture;
 import com.dnd.spaced.global.config.properties.TokenProperties;
 import java.time.LocalDateTime;
